// src/app/shared/interfaces/common.interface.ts
export interface EventData {
     id?: any;
     title: string;
     date: string;
     location: string;
     description: string;
     hashtag?: string;
     image?: string;
     product: string;
     scope: string;
     category: string;
     tags?: any[];
     popularity: number;
     status: string;
     countdown: any;
}